Agregar repositorio de acciones y controlador para gestionar acciones

+++ backend/IntelTask.Domain/Entities/EAcciones.cs
@@ -7,7 +7,7 @@ namespace IntelTask.Domain.Entities
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public byte CN_Id_accion { get; set; }
+        public int CN_Id_accion { get; set; }
         public string? CT_Descripcion_accion { get; set; } = string.Empty;
     }
 }
