+++ backend/IntelTask.Domain/Interfaces/IAccionesRepository.cs
@@ -5,7 +5,7 @@ namespace IntelTask.Domain.Interfaces
     public interface IAccionesRepository
     {
         Task<IEnumerable<EAcciones>> GetAllAccionesAsync();
-        Task<EAcciones?> GetAccionByIdAsync(byte id);
+        Task<EAcciones?> GetAccionByIdAsync(int id);
         Task AddAccionAsync(EAcciones accion);
         Task UpdateAccionAsync(EAcciones accion);
     }
